requirements.txt
asgiref==3.8.1
astroid==3.2.2
dill==0.3.8
Django==5.0
django-restframework==0.0.1
djangorestframework==3.15.2
isort==5.13.2
Markdown==3.6
mccabe==0.7.0
pep8==1.7.1
platformdirs==4.2.2
pylint==3.2.5
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.2
